+++ lib/hydrated_bloc.dart
@@ -1,6 +1,5 @@
 library hydrated_bloc;
 
+export 'package:hydrated_cubit/hydrated_cubit.dart' show HydratedStorage;
+
 export 'src/hydrated_bloc.dart';
-export 'src/hydrated_bloc_delegate.dart';
-export 'src/hydrated_bloc_storage.dart';
-export 'src/hydrated_cipher.dart';