+++ lib/src/hydrated_bloc.dart
@@ -1,42 +1,43 @@
 import 'dart:async';
 
+import 'package:hydrated_cubit/hydrated_cubit.dart';
 import 'package:meta/meta.dart';
 import 'package:bloc/bloc.dart';
 
-import '../hydrated_bloc.dart';
-
 /// {@template hydrated_bloc}
-/// Specialized `Bloc` which handles initializing the `Bloc` state
+/// Specialized [Bloc] which handles initializing the [Bloc] state
 /// based on the persisted state. This allows state to be persisted
 /// across hot restarts as well as complete app restarts.
 /// {@endtemplate}
 abstract class HydratedBloc<Event, State> extends Bloc<Event, State> {
   /// {@macro hydrated_bloc}
-  HydratedBloc() {
-    final stateJson = toJson(state);
+  HydratedBloc(State state) : super(state) {
+    final stateJson = toJson(this.state);
     if (stateJson != null) {
       try {
-        _storage.write(storageToken, stateJson);
+        storage.write(storageToken, stateJson);
       } on dynamic catch (error, stackTrace) {
         onError(error, stackTrace);
       }
     }
   }
 
-  static HydratedBlocDelegate get _delegate =>
-      BlocSupervisor.delegate as HydratedBlocDelegate;
-  static HydratedStorage get _storage => _delegate.storage;
+  /// Instance of [HydratedStorage] which will be used to
+  /// manage persisting/restoring the [Bloc] state.
+  static HydratedStorage storage;
 
-  @mustCallSuper
+  State _state;
+
   @override
-  State get initialState {
+  State get state {
+    if (_state != null) return _state;
     try {
-      final stateJson = _storage.read(storageToken);
-      if (stateJson == null) return null;
+      final stateJson = storage.read(storageToken);
+      if (stateJson == null) return super.state;
       return fromJson(Map<String, dynamic>.from(stateJson));
     } on dynamic catch (error, stackTrace) {
       onError(error, stackTrace);
-      return null;
+      return super.state;
     }
   }
 
@@ -46,11 +47,12 @@ abstract class HydratedBloc<Event, State> extends Bloc<Event, State> {
     final stateJson = toJson(state);
     if (stateJson != null) {
       try {
-        _storage.write(storageToken, stateJson);
+        storage.write(storageToken, stateJson);
       } on dynamic catch (error, stackTrace) {
         onError(error, stackTrace);
       }
     }
+    _state = state;
     super.onTransition(transition);
   }
 
@@ -70,7 +72,7 @@ abstract class HydratedBloc<Event, State> extends Bloc<Event, State> {
   /// `clear` is used to wipe or invalidate the cache of a `HydratedBloc`.
   /// Calling `clear` will delete the cached state of the bloc
   /// but will not modify the current state of the bloc.
-  Future<void> clear() => _storage.delete(storageToken);
+  Future<void> clear() => storage.delete(storageToken);
 
   /// Responsible for converting the `Map<String, dynamic>` representation
   /// of the bloc state into a concrete instance of the bloc state.