+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+# 5.0.0
+
+- **BREAKING**: update to `bloc ^5.0.0`
+- **BREAKING**: extend `hydrated_cubit ^0.1.0`
+- **BREAKING**: `super.initialState` is no longer required
+- docs: minor updates to README
+- docs: logo updates
+
+# 5.0.0-dev.3
+
+- feat: update to `bloc ^5.0.0-dev.11`
+- docs: minor updates to README
+
+# 5.0.0-dev.2
+
+- **BREAKING**: update to `bloc ^5.0.0-dev.10`
+- **BREAKING**: extend `hydrated_cubit ^0.0.3`
+
+# 5.0.0-dev.1
+
+- **BREAKING**: update to `bloc ^5.0.0-dev.7`
+- **BREAKING**: `super.initialState` is no longer required
+
 # 4.1.1
 
 - Remove unnecessary `print` statement